+++ cmd/soroban-rpc/internal/test/get_transactions_test.go
@@ -83,9 +83,9 @@ func TestGetTransactions(t *testing.T) {
 	err := client.CallResult(context.Background(), "getTransactions", request, &result)
 	assert.NoError(t, err)
 	assert.Equal(t, len(result.Transactions), 3)
-	assert.Equal(t, uint32(result.Transactions[0].LedgerSequence), ledgers[0])
-	assert.Equal(t, uint32(result.Transactions[1].LedgerSequence), ledgers[1])
-	assert.Equal(t, uint32(result.Transactions[2].LedgerSequence), ledgers[2])
+	assert.Equal(t, result.Transactions[0].LedgerSequence, ledgers[0])
+	assert.Equal(t, result.Transactions[1].LedgerSequence, ledgers[1])
+	assert.Equal(t, result.Transactions[2].LedgerSequence, ledgers[2])
 
 	// Get transactions from single ledger
 	request = methods.GetTransactionsRequest{
@@ -95,7 +95,7 @@ func TestGetTransactions(t *testing.T) {
 	err = client.CallResult(context.Background(), "getTransactions", request, &result)
 	assert.NoError(t, err)
 	assert.Equal(t, len(result.Transactions), 1)
-	assert.Equal(t, uint32(result.Transactions[0].LedgerSequence), ledgers[0])
+	assert.Equal(t, result.Transactions[0].LedgerSequence, ledgers[0])
 
 	// Get transactions with limit
 	request = methods.GetTransactionsRequest{
@@ -108,7 +108,7 @@ func TestGetTransactions(t *testing.T) {
 	err = client.CallResult(context.Background(), "getTransactions", request, &result)
 	assert.NoError(t, err)
 	assert.Equal(t, len(result.Transactions), 1)
-	assert.Equal(t, uint32(result.Transactions[0].LedgerSequence), ledgers[0])
+	assert.Equal(t, result.Transactions[0].LedgerSequence, ledgers[0])
 
 	// Get transactions using previous result's cursor
 	cursor := result.Pagination.Cursor
@@ -122,7 +122,7 @@ func TestGetTransactions(t *testing.T) {
 	err = client.CallResult(context.Background(), "getTransactions", request, &result)
 	assert.NoError(t, err)
 	assert.Equal(t, len(result.Transactions), 2)
-	assert.Equal(t, uint32(result.Transactions[0].LedgerSequence), ledgers[1])
-	assert.Equal(t, uint32(result.Transactions[1].LedgerSequence), ledgers[2])
+	assert.Equal(t, result.Transactions[0].LedgerSequence, ledgers[1])
+	assert.Equal(t, result.Transactions[1].LedgerSequence, ledgers[2])
 
 }
